Initial checkin of working code
This is a port of the Software Automatic Mouth synthesizer from
https://github.com/s-macke/SAM .

As of this checkin, it compiles and outputs speech to an I2S DAC
properly.  Will be refactored into an AudioGeneratorSAM class in
the coming commits.

Code itself in SAM is ugly as a result of the decompilation utility.