Major rework to 8-bit synthesis, pmem changes
Move to an 8-bit formant synthesizer which reduces the noise on the
output and makes it more intelligible.

Also moved to a real continuous waveform generation instead of a step
one which also reduced noise.

Fixed some signedness issues between INTC and ESP8266 so it generates
identical output to the original(updated repo) code in
https://github.com/earlephilhower/SAM (or the main s-macke repo if
he acceprs the PR).